requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
pymongo==3.10.1
python-dotenv==0.13.0
Werkzeug==1.0.1